+++ src/main/java/com/guigu/system/system/controller/AdminController.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import javax.annotation.Resource;
 import javax.servlet.http.HttpSession;
+import javax.websocket.Session;
 
 import org.omg.PortableServer.ServantActivator;
 import org.springframework.context.support.StaticApplicationContext;
@@ -15,7 +16,9 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.guigu.system.po.Admin;
+import com.guigu.system.po.mapper.PopedomVOMapper;
 import com.guigu.system.system.service.AdminService;
+import com.guigu.system.system.service.PopedomVOService;
 @Controller
 @RequestMapping("/system/admin/")
 public class AdminController {
@@ -23,6 +26,9 @@ public class AdminController {
 	
 	@Resource(name="adminServiceImpl")
 	private AdminService adminService;
+	
+	@Resource(name="popedomVOMapper")
+	private PopedomVOMapper popedomVOMapper;
 	
 	@RequestMapping("list.action")
 	public String list(Admin admin,Model model) {
@@ -68,27 +74,42 @@ public class AdminController {
 	
 	 @RequestMapping("delete.action")
 	    public String delete(Integer adminId,Model model) {
+		 	Admin admin=adminService.findOne(adminId);
+		 	if("是".equals(admin.getAdminRight())) {
+		 		 model.addAttribute("info", "管理员不能被删除");
+		 		  return this.list(null, model);
+		 	}
+		 	int i=popedomVOMapper.findCount(adminId);
+		 	if(i>0) {
+		 		model.addAttribute("info", "该用户与部门有绑定，不能删除！");
+		 		return this.list(null, model);
+		 		
+		 	}
 	        boolean result =adminService.delete(adminId);
-	        
 	        if(result) {
 	            model.addAttribute("info", "删除成功");
-	        }else {
-	            model.addAttribute("info", "删除失败");
 	        }
 	        return this.list(null, model);
 	   }
 	 
 	 @RequestMapping("login.action")
 	 public String login(Admin admin,HttpSession session,Model model) {
-		 admin=adminService.login(admin);
-		 if (admin==null || admin.getAdminState().equals("否")) {
+		 if (admin.getAdminPwd()==null) {
 			model.addAttribute("error", "登录失败");
 			model.addAttribute("admin", admin);
 			return "login";
 		 }
-		 
-		 
-		 
+		 admin=adminService.login(admin);
+		 if (admin==null) {
+				model.addAttribute("error", "登录失败");
+				model.addAttribute("admin", admin);
+				return "login";
+		 }
+		 if(admin.getAdminState().equals("否")) {
+			 model.addAttribute("error", "登录失败");
+			model.addAttribute("admin", admin);
+			return "login";
+		 }
 		 session.setAttribute("admin", admin);  
 		 model.addAttribute("loginadmin", admin);
 		 return "index";
@@ -114,6 +135,12 @@ public class AdminController {
 	        	 model.addAttribute("info", "修改失败");
 			}
 	        return "info/info_main";
+	    }
+	 
+	 @RequestMapping("logout.action")
+	    public String out(HttpSession session) {
+		 	session.invalidate();
+	        return "redirect:/view/login.jsp";
 	    }
 
 }
